+++ src/modules/index.js
@@ -1,6 +1,4 @@
 export { default as toast } from './toast';
-export { default as useWindowDimensions } from './useWindowDimensions';
 export { default as decodeToken } from './decodeToken';
 export { default as logout } from './logout';
-export { default as useToken } from './useToken';
 export { default as history } from './history';
